% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis.createContext}
\alias{analysis.createContext}
\title{Loads the GRC dataset and initializes all the necessary data objects for conducting analyses.}
\usage{
analysis.createContext()
}
\value{
a context object, which is a list containing the necessary data for analyses. This will later be passed to analysis functions.
}
\description{
Loads the GRC dataset and initializes all the necessary data objects for conducting analyses.
}
\examples{
# Not yet done
}
