% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{getGrcFeatureNames}
\alias{getGrcFeatureNames}
\title{Get a list of items that con be measured and mapped}
\usage{
getGrcFeatureNames(grc, category)
}
\arguments{
\item{grc}{The GRC to provides the data to be analyzed (loaded by loadGrc())}

\item{category}{String: The category of features to be listed. Can be one of ["drug", "locus", "mutation", "amplification"]}
}
\value{
A vector of strings, consisting of the names of the  features of the specified category that are supported by this GRC.
}
\description{
The sample features you can analyze depends on the format of each specific GRC. This function can be used
to discover what features can be analyzed for the GRC you have loaded. For example, you can ask for a 
list of "mutation" features, which will return a list of names of mutations that are
supported by the GRC. Then you can supply one or more of these names to the "mutations" parameter of
function mapMutationPrevalence(), to botain prevalence maps.
}
\examples{
\dontrun{
## Load dataset ##
grc <- loadGrc("D:/.../sheet1.xlsx", species="Pf", version="2.0")
mutNames <- getGrcFeatureNames (grc, "mutation")
... 
mapMutationPrevalence(ctx, sampleSet="set1", mutations=mutNames[1:3], aggregate="District")
}
}
