% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{grcMalaria}
\alias{grcMalaria}
\title{grcMalaria: Genetic Epidemiology Analysis using SpotMalaria Genetic Report Cards}
\description{
\if{html}{\figure{logo.jpg}{options: align='right' alt='logo' width='120'}}
A user-friendly, open-source R package, designed to make genetic epidemiology analysis tasks accessible.
This package facilitates the translation of genetic information derived from malaria parasites from SpotMalaria Genetic Report Cards (GRC),
into intuitive geographical maps of prevalence, diversity, relatedness.
This software library is also capable of identifying circulating strains, characterizing drug resistance profiles, and mapping spread.
The R documentation was updated to align with V2.0.0
Refer to the user-guide (link below) for the latest instructions on how to use this package.
}
\section{Author}{

Olivo Miotto \email{olivo@tropmedres.ac}
}

\seealso{
Useful links:
\itemize{
 \item user guide {\url{https://genremekong.org/tools/grcmalaria-r-package-user-guide}}
}
}
