% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{helloWorld}
\alias{helloWorld}
\title{A test function to make sure the package works}
\usage{
helloWorld()
}
\value{
nothing
}
\description{
A test function to make sure the package works
}
\examples{
helloWorld()

}
