% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{initializeContext}
\alias{initializeContext}
\title{Initialize a Genetics Report Cards dataset}
\usage{
initializeContext(
  grcData,
  dir = ".",
  minSnpTypability = 0.8,
  minSampleTypability = 0.75
)
}
\arguments{
\item{grcData}{The data obtained from reading the GRC Excel data.}

\item{dir}{The folder where the outputs from this and subsequent analyses will be stored.}

\item{minSnpTypability}{The minimum proportion of non-missing samples for a barcode position to be retained in analysis. The default is 0.8. Meaning, SNPs where the allele is missing in more than 20\% of samples are removed.}

\item{minSampleTypability}{The minimum proportion of non-missing positions for a sample to be retained in analysis. The default is 0.75. Meaning, samples where more than 25\% of the remaining SNPs are missing are removed.}
}
\value{
A analysis context object, which is a list that contains all the data for analysis, which will be passed to subsequent analysis tasks.
}
\description{
Initialize a GRC dataset (obtained from function loadGrc()) so that it is ready for analysis.
It performs barcode imputation, computes genetic distances, and initializes data that will subsequently be used in analyses.
}
\examples{
#Change the path to where you want output file to be
ctx <- initializeContext(Data,
                         dir="D:/...")
}
