% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{intializeContext}
\alias{intializeContext}
\title{Initialize a Genetics Report Cards dataset
Initialize a GRC dataset (obtained from function loadGrc()) so that it is ready for analysis.
It performs barcode imputation, computes genetic distances, and initializes data that will subsequently be used in analyses.}
\usage{
intializeContext(grcData, dir = ".")
}
\arguments{
\item{grcData}{The data frame obtrained from reading the GRC Excel data.}

\item{dir}{The folder where the outputs from this and subsequent analyses will be stored.}
}
\value{
An analysis context object, which is a list that contains all the data for analysis, which will be passed to subsequent analysis tasks.
}
\description{
Initialize a Genetics Report Cards dataset
Initialize a GRC dataset (obtained from function loadGrc()) so that it is ready for analysis.
It performs barcode imputation, computes genetic distances, and initializes data that will subsequently be used in analyses.
}
\examples{
pfData <- loadGrc ("D:/MyData/GRC/20210705-GenRe.xlsx", sheet="GenRe-Mekong", species="Pf", version="1.0")
ctx <- intializeContext (pfData, dir="D:/MyAnalysis/GRC")
}
