% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapBarcodeFrequencies}
\alias{mapBarcodeFrequencies}
\title{Map barcode group frequencies}
\usage{
mapBarcodeFrequencies(
  ctx,
  sampleSet,
  type = c("bar", "pie"),
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerScale = 0.8,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used; must have been previously created by selectSampleSet().}

\item{type}{Frequency of barcode groups can be visualized either as a "bar" and/or a "pie".}

\item{aggregate}{The administrative level at which we perform pairwise comparisons, either "Province" and/or "District"}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District).}

\item{markerScale}{Allows adjustment of the size of markers on the map, default: 0.8.}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\description{
Creates a map showing the frequency of groups of Plasmodium samples with identical a 'genetic barcode' within administrative divisions.
The 'genetic barcode' consists of 101 SNPs determined for each sample. These SNPs are distributed across all nuclear chromosomes.
SNP positions were chosen based on their ability to differentiate populations and their power to recapitulate genetic distance.
The samples are grouped according to identical barcode within each administrative division, at the desired administrative level: Province, or District.
Barcode groups are not compared between administrative divisions. To compare barcode groups between administrative divisions please use mapClusterSharing(). Use command ?mapClusterSharing for more information.
To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.
Additionally, data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/sampleSetName/map-barcodeFrequency/.
}
\examples{
#Map barcode group frequencies in both bar and pie form, for sampleSet "Laos", on both a Province and District level
mapBarcodeFrequencies (ctx, sampleSet="Laos",
type=c("bar","pie"),
aggregate=c("Province","District"),
minAggregateCount=10, showNames=TRUE, markerScale=0.8,
width=15, height=15)
}
