% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapDiversity}
\alias{mapDiversity}
\title{Title}
\usage{
mapDiversity(
  ctx,
  sampleSet,
  measures = "ALL",
  aggregate = 1,
  minAggregateCount = 10,
  showNames = TRUE,
  markerColours = c("white", "red3"),
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
