% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapGroupFrequencies}
\alias{mapGroupFrequencies}
\title{Map barcode group frequencies}
\usage{
mapGroupFrequencies(
  ctx,
  sampleSet,
  type = c("bar", "pie"),
  similarityLevels = 1,
  aggregate = "Province",
  minGroupSize = 10,
  showNames = TRUE,
  markerScale = 0.8,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{type}{TBD}

\item{similarityLevels}{TBD}

\item{aggregate}{TBD}

\item{minGroupSize}{TBD}

\item{showNames}{TBD}

\item{markerScale}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Map barcode group frequencies
}
\examples{
 #TBD
}
