% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapGroupGraph}
\alias{mapGroupGraph}
\title{Plot barcode similarity graphs}
\usage{
mapGroupGraph(
  ctx,
  sampleSet,
  attributes = NULL,
  similarityLevels = 1,
  minGroupSize = 5,
  minConnectionSimilarity = 0.7,
  graphLayout = "fr",
  weighting = "identitySquared",
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{attributes}{TBD}

\item{similarityLevels}{TBD}

\item{minGroupSize}{TBD}

\item{minConnectionSimilarity}{TBD}

\item{graphLayout}{TBD}

\item{weighting}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Plot barcode similarity graphs
}
\examples{
 #TBD
}
