% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapMutationPrevalence}
\alias{mapMutationPrevalence}
\title{Map the prevalence of genetic mutations.}
\usage{
mapMutationPrevalence(
  ctx,
  sampleSet,
  timePeriods = NULL,
  mutations = "ALL",
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{timePeriods}{The list of time period object for partitioning samples into time-interval plots}

\item{mutations}{TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD}

\item{markerSize}{TBD}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\value{
TBD
}
\description{
Map the prevalence of genetic mutations.
}
\examples{
 #TBD
}
