% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mergeGrc}
\alias{mergeGrc}
\title{Combine two Genetic Report Cards (GRC) datasets}
\usage{
mergeGrc(srcGrc, newGrc, overwrite = FALSE, extendColumns = FALSE)
}
\arguments{
\item{srcGrc}{The source GRC dataset (loaded by loadGrc())}

\item{newGrc}{The new GRC dataset to be added(loaded by loadGrc())}

\item{overwrite}{Boolean: if TRUE, sample rows in newGrc will replace rows in srcGrc that have the same sampleId. If FALSE, they will be disregarded.}

\item{extendColumns}{Boolean: if TRUE, new columns in newGrc will also be added to srcGrc, filled with <NA> values.  If FALSE, they will be disregarded.}
}
\value{
A list containing a data frame with the merged data ready to be analyzed, plus some configuration metadata
}
\description{
Merge two GRC datasets which have been loaded from data files. The new dataset is added to the source dataset. 
If the new dataset contains new columns, they can be either ignored, or the same columns are added to the source dataset, 
with null values. If the new dataset contains sample data that is already in the source dataset, they can be either ignored, 
or the source data rows can be overwritten.
}
\examples{
\dontrun{
# Load spreadsheet 1
Sheet1 <- loadGrc("D:/myFiles/sheet1.xlsx",
                 sheet="GRC", species="Pf", version="1.2")
# Load spreadsheet 2
Sheet2 <- loadGrc("D:/myFiles/sheet2.xlsx",
                 sheet="GRC", species="Pf", version="1.2")
# Merge datasets
mergeData <- mergeGrc(Sheet1, Sheet2)
}
}
