% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotBarcodeNetwork}
\alias{plotBarcodeNetwork}
\title{Plot Population structure (barcode networks)}
\usage{
plotBarcodeNetwork(
  ctx,
  sampleSet,
  attributes = NULL,
  aggregateCountMin = 10,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{attributes}{TBD}

\item{aggregateCountMin}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Plot Population structure (barcode networks)
}
\examples{
 #TBD
}
