% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotClusterGraph}
\alias{plotClusterGraph}
\title{Plot graphs of bracode identity, using cluster}
\usage{
plotClusterGraph(
  ctx,
  sampleSet,
  clusterSet,
  graphLayout = "fr",
  weightPower = 2,
  ...
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{clusterSet}{TBD}

\item{graphLayout}{TBD}

\item{weightPower}{TBD}

\item{...}{any of plot paramenters, including: width, height, units, dpi, legendPosition, legendWidth}
}
\value{
TBD
}
\description{
Plot graphs of bracode identity, using cluster
}
\examples{
 #TBD
}
