% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot Population structure (PCA)}
\usage{
plotPCA(
  ctx,
  sampleSet,
  method = "PCoA",
  attributes = NULL,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD TBD}

\item{method}{TBD}

\item{attributes}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Plot Population structure (PCA)
}
\examples{
 #TBD
}
