% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotPopulationStructure}
\alias{plotPopulationStructure}
\title{Title}
\usage{
plotPopulationStructure(
  ctx,
  sampleSet,
  type = c("njt", "PCoA"),
  attributes = NULL,
  width = 15,
  height = 15
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
