% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotPrincipalComponents}
\alias{plotPrincipalComponents}
\title{Plot Population structure (PCA)}
\usage{
plotPrincipalComponents(
  ctx,
  sampleSet,
  type = "PCoA",
  plots,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD TBD}

\item{type}{TBD}

\item{plots}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Plot Population structure (PCA)
}
\examples{
 #TBD
}
