% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotTree}
\alias{plotTree}
\title{Plot Trees}
\usage{
plotTree(ctx, sampleSet, type = "njt", plots, ...)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD TBD}

\item{type}{TBD}

\item{plots}{TBD}

\item{...}{any of plot paramenters, including: width, height, units, dpi, legendPosition, legendWidth}
}
\value{
TBD
}
\description{
Plot Trees
}
\examples{
 #TBD
}
