% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{resetColourPalette}
\alias{resetColourPalette}
\title{Reset to the default colour palette, removing user-defined palettes}
\usage{
resetColourPalette(ctx)
}
\arguments{
\item{ctx}{TBD}
}
\value{
TBD
}
\description{
Reset to the default colour palette, removing user-defined palettes
}
\examples{
 #TBD
}
