% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{setColourPalette}
\alias{setColourPalette}
\title{Set to a user-defined colour palette}
\usage{
setColourPalette(ctx, palette)
}
\arguments{
\item{ctx}{TBD}

\item{palette}{TBD}
}
\value{
TBD
}
\description{
Set to a user-defined colour palette
}
\examples{
 #TBD
}
