% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{findClusters}
\alias{findClusters}
\title{Clustering}
\usage{
findClusters(
  ctx,
  sampleSet,
  clusterSet,
  minIdentity = 1,
  impute = TRUE,
  clusteringMethod = "allNeighbours",
  minClusterSize = 10
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{clusterSet}{TBD}

\item{minIdentity}{TBD}

\item{impute}{TBD}

\item{clusteringMethod}{TBD}

\item{minClusterSize}{TBD}
}
\value{
TBD
}
\description{
Clustering
}
\examples{
 #TBD
}
