% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\docType{package}
\name{grcMalaria}
\alias{grcMalaria}
\title{grcMalaria: Processing and analyzing SpotMalaria Genetic Report Cards}
\description{
\if{html}{\figure{logo.jpg}{options: align='right' alt='logo' width='120'}}
Various functions for producing analysis results such as maps and reports from the Genetic Report Cards datasets
delivered by SpotMalaria projects, such as GenRe-Mekong.
Typical outputs include drug resistance prevalence maps, population genetic analyses such as PCA (principal component analysis) and NJ (neighbor joining) trees,
and analyses of genetic barcodes such as barcode sharing networks, etc.
The data frame returned contains a data table whose values and structure may have been manipulated after being read from the file,
and may have been updated to the latest version fo the format.
}
\section{Author}{

Olivo Miotto \email{olivo@tropmedres.ac}
}

\seealso{
Useful links:
\itemize{
 \item user guide {\url{https://cutt.ly/grcMalaria-R-UserGuide}}
 \item source code {\url{https://github.com/malariagen/grcMalaria}}
 \item reporting bugs {\url{https://github.com/malariagen/grcMalaria/issues}}
}
}
