% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{loadGraphicAttributes}
\alias{loadGraphicAttributes}
\title{Graphical Attributes for principal component analysis and neighbour-joining tree}
\usage{
loadGraphicAttributes(ctx, name, field, file, sheet)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{name}{The name of the graphical attribute, this can be any metadata category of interest.
Once defined, graphical attributes can be applied by referencing their name.
Examples of names are: 'Countries', 'Provinces', 'Kelch13', etc.}

\item{field}{The corresponding column-name in the original data file of the specified graphical attributes.}

\item{file}{The file name of the Excel spreadsheet where all graphical attributes are specified.}

\item{sheet}{The sheet name of the Excel spreadsheet where the to parameter 'name' corresponding graphical attribute is specified.}
}
\description{
This function allows for specification of graphical display of samples based on their metadata,
that can subsequently be used in a principal component analysis (plotPrincipalComponents()),
or creation of a neighbour-joining tree (plotTree()).
Give command ?plotPrincipalComponents, or ?plotTree for more information.
Specification based on metadata (i.e. attributes) is done by loading an Excel spreadsheet into R.
The spreadsheet and instructions can be found in the grcMalaria user-guide, see link in 'See also' section.
}
\examples{
 #Load Excel spreadsheet into R with specified graphical attributes
 gaFile <- "/path/to/my/file/GenRe-GraphicAttributes.xlsx"

 #Load graphical attributes for province, Kelch13, and Plasmepsin 2,3
ctx <- loadGraphicAttributes (ctx, name="province",      field="AdmDiv1",       file=gaFile, sheet="Provinces")
ctx <- loadGraphicAttributes (ctx, name="k13",           field="Kelch",         file=gaFile, sheet="Kelch13")
ctx <- loadGraphicAttributes (ctx, name="pm23",          field="Plasmepsin2/3", file=gaFile, sheet="Plasmepsin23")
ctx <- loadGraphicAttributes (ctx, name="pm23-noColour", field="Plasmepsin2/3", file=gaFile, sheet="Plasmepsin23-noColour")

}
\seealso{
grcMalaria user-guide:
\itemize{
 \item {\url{https://olivo-miotto-genre.notion.site/grcMalaria-R-package-User-Guide-dfaac7ea30b5430b8522e8b531fd2e4b}}
}
}
