% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{loadGraphicAttributes}
\alias{loadGraphicAttributes}
\title{Graphical Attributes}
\usage{
loadGraphicAttributes(ctx, name, field, file, sheet)
}
\arguments{
\item{ctx}{TBD}

\item{name}{TBD}

\item{field}{TBD}

\item{file}{TBD}

\item{sheet}{TBD}
}
\value{
TBD
}
\description{
Graphical Attributes
}
\examples{
 #TBD
}
