% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{loadGrc}
\alias{loadGrc}
\title{Load a Genetic Report Cards data file
Load a GRC data file, ready for analysis.
The data frame returned constaind a data table whose values and structure may have been manipulated after being read from the file,
and may have been updated to the latest version fo the format.}
\usage{
loadGrc(file, sheet = "GenRe-Mekong", species = "Pf", version = "1.0")
}
\arguments{
\item{file}{The path to the GRC data Microsoft Excel file (use forward slashes in the path)}

\item{sheet}{The name of the datasheet within the Excel file containing the GRC data}

\item{species}{The species being analyzed ("Pf" or "Pv")}

\item{version}{The version number of the GRC data file format. This is in the documentation when you receive or download the file.}
}
\value{
a data frame containing the data ready to be analyzed
}
\description{
Load a Genetic Report Cards data file
Load a GRC data file, ready for analysis.
The data frame returned constaind a data table whose values and structure may have been manipulated after being read from the file,
and may have been updated to the latest version fo the format.
}
\examples{
pfData <- loadGrc ("D:/MyData/GRC/20210705-GenRe.xlsx", sheet="GenRe-Mekong", species="Pf", version="1.0")
}
