% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapAlleleProportions}
\alias{mapAlleleProportions}
\title{Map the proportion of alleles for a given mutation (including kelch13 and amplifications).}
\usage{
mapAlleleProportions(
  ctx,
  sampleSet,
  timePeriods = NULL,
  mutations = "ALL",
  aggregate = "Province",
  minAggregateCount = 10,
  markerSize = 16,
  showNames = TRUE,
  nameFontSize = 5,
  ...
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext()}

\item{sampleSet}{The name of the sample set being used; must have been previously created by selectSampleSet()}

\item{timePeriods}{A separately given list of time period objects for partitioning samples into time-interval plots. Each list specifies a time period, with three parameters: name, type and, start. 
Parameter type can currently only be year. Parameter start must be a date in the format ?dd-MMM-yyyy?, and parameter name must be provided. The string passed in the name parameter is added to the end of the file name of the produced files.
Currently, the public GRC data only contains years, so only the year in the provided ?start? date is used}

\item{mutations}{The genetic mutation types for which prevalence will be calculated; "ALL" creates maps for all 11 available genetic mutations
Available genetic mutations include: "Pfkelch13", "pm23-Amp", "mdr1-Amp", "crt", "dhfr", "dhps", "mdr1", "mdr2", "arps10", "fd", "exo"
To specify a mutation put it in between quotation marks e.g. "Pfkelch13", or  c("Pfkelch13","dhfr","fd") to select several genetic mutation types}

\item{aggregate}{The administrative level at which we aggregate}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown}

\item{markerSize}{Allows adjustment of the size of markers on the map}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District)}

\item{nameFontSize}{Set the font size of the geographical name labels outside the markers on the map, if showNames=TRUE, default=5}

\item{...}{any of plot parameters, including: width, height, units, dpi, legendPosition. 
width: the width of the plot (numeric), default=15
height: the height of the plot (numeric), default=15
units: the units in which the width and height are expressed. Supported values: "in" (default), "cm", "mm", "px"
dpi: the resolution of the plot output, expressed as dots per inch, default=300
format: the file format in which the plot will be saved. Supported values: "png" (default), "pdf"
legendPosition: specifies where the legend should be plotted. Supported values: "inset" (default), "separate"}
}
\description{
Creates maps showing the proportions of observed allele variants of a given genetic mutation.
If multiple genetic mutations are specified, then a map will be created for each mutation.
Missing genotypes and heterozygous genotypes are excluded in the calculations.
The proportions of observed allele variants are aggregated at the desired administrative level: Province (level 1), or District (level 2). Separate maps are created for each allele-administrative-combination.
For each aggregation unit, a pie-chart is created, coloured according to the observed proportions of observed allele variants of the mutation of interest.
To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.
Additionally data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/sampleSetName/map-alleleProp/.
}
\examples{
\dontrun{
# Produce maps for all available genetic mutations for sample set called Cambodia, 
# at province level, with a minimum of 10 samples
mapAlleleProportions (ctx, sampleSet="Cambodia", 
                      mutations="ALL",
                      aggregate="Province", minAggregateCount=10,
                      showNames=TRUE, markerSize=c(10,25))
}
}
\seealso{
Useful links:
\itemize{
 \item user guide for information on functionality of specific genetic mutation types {\url{https://cutt.ly/grcMalaria-R-UserGuide}}
}
}
