% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapBarcodeFrequencies}
\alias{mapBarcodeFrequencies}
\title{Map barcode group frequencies}
\usage{
mapBarcodeFrequencies(
  ctx,
  sampleSet,
  type = c("bar", "pie"),
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerScale = 0.8,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{type}{TBD}

\item{aggregate}{TBD}

\item{showNames}{TBD}

\item{markerScale}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Map barcode group frequencies
}
\examples{
 #TBD
}
