% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapClusterPrevalence}
\alias{mapClusterPrevalence}
\title{Map Cluster prevalence at different sites, connecting these sites}
\usage{
mapClusterPrevalence(
  ctx,
  sampleSet,
  clusterSet,
  aggregate = "Province",
  minAggregateCount = 5,
  showNames = TRUE,
  markerScale = 0.8,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{clusterSet}{TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD}

\item{markerScale}{TBD}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\value{
TBD
}
\description{
Map Cluster prevalence at different sites, connecting these sites
}
\examples{
 #TBD
}
