% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapClusterPrevalence}
\alias{mapClusterPrevalence}
\title{Map cluster prevalence at different sites, connecting these sites}
\usage{
mapClusterPrevalence(
  ctx,
  sampleSet,
  clusterSet,
  aggregate = "Province",
  minAggregateCount = 5,
  markerSize = 16,
  markerFontSize = 6,
  showNames = TRUE,
  nameFontSize = 5,
  ...
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used, which must have been previously created by selectSampleSet().}

\item{clusterSet}{The name of the clustering set, defined in findClusters().}

\item{aggregate}{The administrative level at which we perform pairwise comparisons, either "Province" and/or "District".}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples,
below which the marker is not shown, default=5}

\item{markerSize}{Allows adjustment of the size of markers on the map.}

\item{markerFontSize}{Set the font size of the value labels inside the markers on the map, default=6}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District)}

\item{nameFontSize}{Set the font size of the geographical name labels outside the markers on the map, if showNames=TRUE, default=5}

\item{...}{any of plot parameters, including: width, height, units, dpi, legendPosition
width: the width of the plot (numeric), default=15
height: the height of the plot (numeric), default=15
units: the units in which the width and height are expressed. Supported values: "in" (default), "cm", "mm", "px"
dpi: the resolution of the plot output, expressed as dots per inch, default=300
format: the file format in which the plot will be saved. Supported values: "png" (default), "pdf"
legendPosition: specifies where the legend should be plotted. Supported values: "inset" (default), "separate"}
}
\value{
For each assigned cluster a map with cluster prevalence and overlap in cluster prevalence across administrative divisions
}
\description{
For each assigned cluster a map is created, showing the prevalence and overlap in prevalence of that cluster across administrative divisions.
Each map also contains information on sample count, drug-resistance, and mutation frequencies.
The maps are based on the cluster analysis, which needs to be produced with the findClusters function in advance.
The findClusters function partitions the Plasmodium samples into clusters with a similar genetic background, based on the genetic barcode similarity.
Use command ?findClusters for more information.
The genetic barcode consists of 101 SNPs determined for each sample. These SNPs are distributed across all nuclear chromosomes.
SNP positions were chosen based on their ability to differentiate populations and their power to recapitulate genetic distance.
The prevalence of the assessed cluster is displayed in the nodes.
The overlap in prevalence is displayed in the edges (i.e. lines).
The definition of prevalence overlap is the proportion of sample pairs that have at least the assigned "minIdentity" of genetic barcode similarity in the findCluster function.
Additionally, data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/(sampleSetName)/map-clusterPrevalence/.
}
\examples{
\dontrun{
# Produce a network and prevalence map for each assigned cluster,
# for sampleSet "Laos", and clusterSet "LAclust" at both a Province and District level,
mapClusterPrevalence (ctx, sampleSet="Laos", clusterSet = "LAclust",
                      aggregate=c("Province","District"),
                      minAggregateCount=5, showNames=TRUE)
}
}
