% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapClusterPrevalence}
\alias{mapClusterPrevalence}
\title{Map cluster networks between sites based on prevalence}
\usage{
mapClusterPrevalence(
  ctx_cluster,
  sampleSet,
  clusterSet,
  aggregate = "Province",
  minAggregateCount = 5,
  showNames = TRUE,
  markerScale = 0.8,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx_cluster}{The cluster analysis context, created by findClusters().}

\item{sampleSet}{The name of the sample set being used; must have been previously created by selectSampleSet().}

\item{clusterSet}{The name of the clustering set, defined in findClusters().}

\item{aggregate}{The administrative level at which we perform pairwise comparisons, either "Province" and/or "District". The default is "Province".}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples,
below which the marker is not shown. The default is 5.}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District).}

\item{markerScale}{Allows adjustment of the size of markers on the map. The default is 0.8.}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\value{
For each assigned cluster a map with cluster prevalence and overlap in cluster prevalence across administrative divisions
}
\description{
For each assigned cluster a map is created, showing the prevalence and overlap in prevalence of that cluster across administrative divisions.
Each map also contains information on sample count, drug-resistant and mutation frequencies.
The maps are based on the cluster analysis context, which needs to be produced with the findClusters function in advance.
The findClusters function partitions the Plasmodium samples into clusters with a similar genetic background, based on the 'genetic barcode' similarity.
Use command ?findClusters for more information.
The 'genetic barcode' consists of 101 SNPs determined for each sample. These SNPs are distributed across all nuclear chromosomes.
SNP positions were chosen based on their ability to differentiate populations and their power to recapitulate genetic distance.
The prevalence of the assessed cluster is displayed in the nodes.
The overlap in prevalence is displayed in the edges (i.e. lines).
The definition of prevalence overlap is the proportion of sample pairs that have at least the assigned "minIdentity" of genetic barcode similarity in the findCluster function.
Additionally, data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/(sampleSetName)/map-clusterPrevalence/.
}
\examples{
# Produce a network and prevalence map for each assigned cluster,
# using cluster analysis context called ctx_cluster,
# for sampleSet "Laos", and clusterSet "LAclust" on both a Province and District level,
mapClusterPrevalence (ctx_cluster, sampleSet="Laos", clusterSet = "LAclust",
aggregate=c("Province","District"),
minAggregateCount=5, showNames=TRUE, markerScale=0.8,
width=15, height=15)
}
