% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapConnections}
\alias{mapConnections}
\title{Map Connectedness between sites}
\usage{
mapConnections(
  ctx,
  sampleSet,
  measures = "ALL",
  minIdentity = 1,
  meanDistanceLevels = 0.5,
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used; must have been previously created by selectSampleSet().}

\item{measures}{default is "ALL", other options are "meanDistance", or "similarity".
"meanDistance" produces a map with pairwise comparisons between administrative divisions (e.g. sites) in which the thickness of the connection line corresponds to
the proportion of sample pairs with a mean genetic distance of at least the meanDistanceLevels
"similarity" produces a map with pairwise comparisons between administrative divisions (e.g. sites) in which the thickness of the connection line corresponds to
the proportion of sample pairs with at least the percentage of similarity given in minIdentity.}

\item{minIdentity}{value between 0 and 1, default is 1}

\item{meanDistanceLevels}{value between 0 and 1, default is 0.5}

\item{aggregate}{The administrative level at which we perform pairwise comparisons}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District).}

\item{markerSize}{Allows adjustment of the size of markers on the map.}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\description{
Creates a map of connectedness of haplotypes between different administrative divisions (i.e. sites) in the selected sampleSet.
The administrative divisions are compared in a pairwise fashion, within this comparison output is produced that
gives either i) the proportion of sample pairs that have at least the assigned percentage of similarity (default: 1), or
ii) the proportion of sample pairs with a mean genetic distance of at least the meanDistanceLevels (default: 0.5).
Subsequently, maps are produced that show connections between the administrative divisions,
in which the thickness of the connection line corresponds to the proportion of either i), or ii).
Both similarity and mean genetic distance are calculated based on the 'genetic barcode', which consists of 101 SNPs determined for each sample.
These SNPs are distributed across all nuclear chromosomes.
SNP positions were chosen based on their ability to differentiate populations and their power to recapitulate genetic distance.
If multiple levels are specified at minIdentity, and meanDistanceLevels, then multiple maps will be produced.
The connectedness is aggregated at the desired administrative level: Province (level 1), or District (level 2).
To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.
Additionally data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/sampleSetName/map-connect/.
}
\examples{
#Map connectedness of haplotypes between districts for a similarity of 1.0, and 0.95, and a mean genetic distance of 0.5, 0.6, and 0.7
mapConnections (ctx, sampleSet="Laos",
measures="ALL", aggregate="District", minIdentity=c(1.0,0.95),
meanDistanceLevels=c(0.5,0.6,0.7),
minAggregateCount=10, showNames=TRUE,
width=15, height=15)
}
