% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapConnections}
\alias{mapConnections}
\title{Map Connectedness between sites}
\usage{
mapConnections(
  ctx,
  sampleSet,
  measures = "ALL",
  similarityLevels = 1,
  meanDistanceLevels = 0.5,
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{measures}{TBD}

\item{similarityLevels}{TBD TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD}

\item{markerSize}{TBD}

\item{width}{TBD}

\item{height}{TBD}

\item{markerColours}{TBD}
}
\value{
TBD
}
\description{
Map Connectedness between sites
}
\examples{
 #TBD
}
