% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapConnections}
\alias{mapConnections}
\title{Map Connectedness between sites}
\usage{
mapConnections(
  ctx,
  sampleSet,
  measures = "ALL",
  minIdentity = 1,
  meanDistanceLevels = 0.5,
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  ...
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{measures}{TBD}

\item{minIdentity}{TBD}

\item{meanDistanceLevels}{TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD}

\item{markerSize}{TBD}

\item{...}{any of plot paramenters, including: width, height, units, dpi, legendPosition, legendWidth}
}
\value{
TBD
}
\description{
Map Connectedness between sites
}
\examples{
 #TBD
}
