% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapDiversity}
\alias{mapDiversity}
\title{Map genetic Diversity}
\usage{
mapDiversity(
  ctx,
  sampleSet,
  timePeriods = NULL,
  measures = "ALL",
  aggregate = "Province",
  minAggregateCount = 10,
  markerSize = 16,
  markerFontSize = 6,
  markerColours = "red3",
  showNames = TRUE,
  nameFontSize = 5,
  ...
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used, which must have been previously created by selectSampleSet().}

\item{timePeriods}{Time-sequence maps can be implemented using a parameter called timePeriods parameter. When this is passed to a function, it will partition samples into time-interval plots. Time intervals parameters are available in 5 analyses: mapSampleCounts(), mapDrugResistancePrevalence(), mapMutationPrevalence(), mapAlleleProportions() and mapDiversity().
Why use time-intervals parameter:
It will produce maps with consistent geographical boundaries
Users can slice up the dataset in any specified time period}

\item{measures}{This can be "ALL", or any vector containing one or more of c("maxBarcodeFreq","barcodeHet", "meanVariantHet", "meanHetCallsProp", "medianDistance").
The method "maxBarcodeFreq" is a measure of loss of diversity. This measure gives the proportion of samples carrying the most common 
barcode haplotype (defined as samples with identical barcodes). The output ranges from 0-1, with a low value corresponding to a low 
proportion of samples with the most common haplotype (high diversity), and a high value corresponding to a high proportion of samples 
with the most same haplotype (low diversity).
The method "barcodeHet" is a measure of heterozygosity based on the complete barcode. This measure is the probability of two randomly 
selected samples carrying different barcodes, and is useful to detect clonal expansions.
The output ranges from 0-1, with a low value corresponding to low diversity (low probability of carrying a different barcode), 
while a high value corresponds to high diversity (high probability of a different barcode).
The method "meanVariantHet" is the average of the expected heterozygosity of the individual variants in a barcode.  For each barcode SNP, 
the expected heterozygosity is calculated using: HE=(n/(n-1))(1-sum(pi^2)), where n = the number of samples and pi = the allele frequency 
of the ith allele at this SNP. The final diversity estimate is the mean of SNP heterozygosity across all barcoding SNPs.
The output ranges from 0-1, with a low value corresponding to low diversity (low probability of a different allele), and a high value 
corresponding to high diversity (high probability of a different allele).
The method "meanHetCallsProp" is the mean of the proportion of heterozygous calls at each of the barcoding SNPs. For each barcode SNP, 
we count how many samples carry more than one allele at this SNP, and estimate the proportion, ignoring those samples that have a 
missing genotype. The final diversity estimate is the mean these proportions, across all barcoding SNPs.
The output ranges from 0-1, with a low value corresponding to low diversity, and a high value corresponding to high diversity.
The method "medianDistance" is a measure that gives the median genetic distance in geographical unit.
It is calculated using pairwise genetic distance as the proportion of SNPs differing between two samples, then taking the median 
of these in the group of samples of interest
The method "meanDistance" is a measure that gives the mean genetic distance in geographical unit, with the same method as medianDistance.}

\item{aggregate}{The administrative level at which we aggregate. Separate maps are created for each administrative level.}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.}

\item{markerSize}{Allows adjustment of the size of markers on the map. If only one value is passed, all markers will be of that size; if two values are passed, they will be used as the min and max size of the marker, whose size will reflect the number of samples.}

\item{markerFontSize}{Allows adjustment of the font size shown on the markers (numeric), default=6.}

\item{markerColours}{The colour to indicate the level of genetic diversity. Default: "red3", other examples: "forestgreen", "cornflowerblue","darkgoldenrod3"}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District)}

\item{nameFontSize}{Set the font size of the geographical name labels outside the markers on the map, if showNames=TRUE, default=5}

\item{...}{Aesthetics: any of the following plot parameters,
width: the width of the plot (numeric), default=15
height: the height of the plot (numeric), default=15
units: the units in which the width and height are expressed. Supported values: "in" (default), "cm", "mm", "px"
dpi: the resolution of the plot output, expressed as dots per inch, default=300
format: the file format in which the plot will be saved. Supported values: "png" (default), "pdf"
legendPosition: specifies where the legend should be plotted. Supported values: "inset" (default), "separate"
legendWidth: specifies how wide a fixed width legend space should be, default="NULL"
legendDirection: specifies location of legend. Supported values: "vertical" (default), "horizontal"
legendFontSize: specifies font size of the legend (numeric), default=4
axisTitleSize: specifies axis label font size (numeric), default=1}
}
\description{
This creates maps of expected heterozygosity or genetic diversity for different administrative divisions.
Genetic diversity can be calculated using four different measures: "maxBarcodeFreq","barcodeHet", "meanVariantHet", and "medianDistance".
Calculations are based on the 'genetic barcode', which consists of 101 SNPs determined for each sample. These SNPs are distributed across all nuclear chromosomes.
SNP positions were chosen based on their ability to differentiate populations and their power to recapitulate genetic distance.
If multiple measures are specified, then a map will be created for each measure.
For each aggregation unit, we place a marker on the map, coloured according to the genetic diversity, with a label indicating the genetic diversity.
}
\examples{
\dontrun{
#' #Given lists of time periods of interest, in this case calendar years
periods <- list(
               list(name="2018", type="year", start="1-Jan-2018"),
               list(name="2019", type="year", start="1-Jan-2019"),
               list(name="2020", type="year", start="1-Jan-2020")
           )

# Produce genetic diversity maps for all available measures, for sampleSet "Laos",
# at District level, and for calendar years 2018-2020.
mapDiversity (ctx, sampleSet="Laos", timePeriods=periods,
              measures="ALL", aggregate="District", markerColours="red3",
              minAggregateCount=10, showNames=TRUE, markerSize=16,
              width=20, height=20)
}
}
