% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapDiversity}
\alias{mapDiversity}
\title{Map genetic Diversity}
\usage{
mapDiversity(
  ctx,
  sampleSet,
  measures = "ALL",
  aggregate = 1,
  minAggregateCount = 10,
  showNames = TRUE,
  markerColours = c("white", "red3"),
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{measures}{TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD}

\item{markerColours}{TBD}

\item{markerSize}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Map genetic Diversity
}
\examples{
 #TBD
}
