% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapDrugResistancePrevalence}
\alias{mapDrugResistancePrevalence}
\title{Map prevalence of Drug resistance (Predicted Phenotype)}
\usage{
mapDrugResistancePrevalence(
  ctx,
  sampleSet,
  timePeriods = NULL,
  drugs = "ALL",
  aggregate = "Province",
  minAggregateCount = 10,
  markerSize = 16,
  markerFontSize = 6,
  showNames = TRUE,
  nameFontSize = 5,
  ...
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext()}

\item{sampleSet}{The name of the sample set being used, which must have been previously created by selectSampleSet()}

\item{timePeriods}{Time-sequence maps can be implemented using a parameter called timePeriods parameter. When this is passed to a function, it will partition samples into time-interval plots. Time intervals parameters are available in 5 analyses: mapSampleCounts(), mapDrugResistancePrevalence(), mapMutationPrevalence(), mapAlleleProportions() and mapDiversity().
Why use time-intervals parameter:
It will produce maps with consistent geographical boundaries
Users can slice up the dataset in any specified time period}

\item{drugs}{The antimalarial drugs for which prevalence of phenotypic resistance will be estimated; "ALL" creates maps for all the drugs for which phenotypic resistance predictions are available,
which include "Artemisinin", "Chloroquine", "Piperaquine", "DHA-PPQ" (i.e. Dihydroartemisinin/piperaquine), "Mefloquine", "Sulfadoxine", "Pyrimethamine", "S-P" (i.e. Sulphadoxine-Pyrimethamine),
"AS-MQ" (i.e. Artesunate-Melfoquine), and "S-P-IPTp" (i.e. intermittent preventive treatment of malaria in pregnancy using Sulfadoxine-Pyrimethamine)
To specify a drug put the drug name in between quotation marks e.g. "Artemisinin", or  c("Artemisinin", "Chloroquine", "S-P") to select several specific drugs.}

\item{aggregate}{The administrative level at which we aggregate. Separate maps are created for each administrative level.}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.}

\item{markerSize}{Allows adjustment of the size of markers on the map. If only one value is passed, all markers will be of that size; if two values are passed, they will be used as the min and max size of the marker, whose size will reflect the number of samples.}

\item{markerFontSize}{Allows adjustment of the font size shown on the markers (numeric), default=6.}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District).}

\item{nameFontSize}{Allows adjustment of the Province or District label name font size (numeric), default=5.}

\item{...}{Aesthetics: any of the following plot parameters,
width: the width of the plot (numeric), default=15
height: the height of the plot (numeric), default=15
units: the units in which the width and height are expressed. Supported values: "in" (default), "cm", "mm", "px"
dpi: the resolution of the plot output, expressed as dots per inch, default=300
format: the file format in which the plot will be saved. Supported values: "png" (default), "pdf"
legendPosition: specifies where the legend should be plotted. Supported values: "inset" (default), "separate"
legendWidth: specifies how wide a fixed width legend space should be, default="NULL"
legendDirection: specifies location of legend. Supported values: "vertical" (default), "horizontal"
legendFontSize: specifies font size of the legend (numeric), default=4
axisTitleSize: specifies axis label font size (numeric), default=1}
}
\description{
Creates a map showing the levels of resistance to a particular antimalarial drug for different administrative divisions.
Based on published genetic markers, each sample has a predicted phenotype for different types of antimalarial drugs.
If multiple drugs are specified, then different maps will be created for different drugs.
}
\examples{
\dontrun{
## To implement time-sequence maps ##
# 1. Specify time periods
# name: the name parameter is required and is added to the end of the file name of the produced 
# files.
# There are two types of time interval: type="year" uses start date and by default show until 
# the end of the defined year; type="period" allows user to specify any time period using both 
# the start and end date.
# start and end - start and end date. The parameter start must be a date in the 
# format dd-MMM-yyyy.
# NOTE: Public GRC data files contains only year of collection, therefore only the year in the 
# provided “start” date is used to produce time-sequence maps.
periods <- list(
               list(name="2021", type="year", start="1-Jan-2021"),
               list(name="2020", type="year", start="1-Jan-2020"),
               list(name="2017-19", type="period", start="1-Jan-2017", end="31-Dec-2019")
            )

# 2. Apply timePeriods parameter to mapDrugResistancePrevalence
mapDrugResistancePrevalence(ctx, sampleSet="Laos", timePeriods = periods,
                            drugs="ALL", aggregate=c("Province","District"),
                            minAggregateCount=10, showNames=TRUE, markerSize=16,
                            width=15, height=15)

# How to read the output:
# Number on the marker shows the proportion of samples with resistant markers for the specified 
# drug, ranging between 0 and 1 where 0 means no samples carry the resistant marker and 1 means 
# 100\% of the samples carry the markers used to predict resistance.
# Resistance prevalence is calculated by dividing the number of resistant samples by the sum of 
# resistant and sensitive samples, fr = Resistant samples/(Resistant + Sensitive sample)
}

}
\seealso{
Useful links:
\itemize{
\item Details of how drug resistance phenotypes are predicted based on genetic markers {\url{http://ngs.sanger.ac.uk/production/malaria/Resource/29/20200705-GenRe-05-PhenotypeRules-0.39.pdf}}
\item Other useful background information about the SpotMalaria platform  {\url{https://www.malariagen.net/resource/29}}
}
}
