% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapDrugResistancePrevalence}
\alias{mapDrugResistancePrevalence}
\title{Map prevalence of Drug resistance (predicted phenotype)}
\usage{
mapDrugResistancePrevalence(
  ctx,
  sampleSet,
  timePeriods = NULL,
  drugs = "ALL",
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used; must have been previously created by selectSampleSet().}

\item{timePeriods}{A separately given list of time period object for partitioning samples into time-interval plots. Each list specifies a time period, with three parameters: name, type and, start. Parameter type can currently only be year. Parameter start must be a date in the format “dd-MMM-yyyy”, and parameter name must be provided. The string passed in the name parameter is added to the end of the file name of the produced files.
Currently, the public GRC data only contains years, so only the year in the provided “start” date is used.}

\item{drugs}{The antimalarial drugs for which prevalence of phenotypic resistance will be estimated; "ALL" creates maps for all the drugs for which phenotypic resistance predictions are available,
which include "Artemisinin", "Chloroquine", "Piperaquine", "DHA-PPQ" (i.e. Dihydroartemisinin/piperaquine), "Chloroquine", and "Pyrimethamine".
To specify a drug put the drug name in between quotation marks e.g. "Artemisinin", or  c(“Artemisinin”, “Chloroquine”, “S-P”) to select several specific drugs.}

\item{aggregate}{The administrative level at which we aggregate}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District)}

\item{markerSize}{Allows adjustment of the size of markers on the map.}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\description{
Creates a map showing the levels of estimated resistance to a particular antimalarial drug. Based on published genetic markers, each sample has a predicted phenotype for different types of antimalarial drugs.
If multiple drugs are specified, then different maps will be created for different drugs.
The predictions of resistance for any given drug are aggregated at the desired administrative level: Province (level 1), or District (level 2). Separate maps are created for each drug-administrative-combination.
For each aggregation unit, we place a marker on the map, coloured according to the level of resistance to the drug, with a label indicating the prevalence.
To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.
Additionally data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/sampleSetName/map-drug/.
}
\examples{
#Given lists of time periods of interest, in this case calender years
periods <- list(
list(name="2018", type="year", start="1-Jan-2018"),
list(name="2019", type="year", start="1-Jan-2019"),
list(name="2020", type="year", start="1-Jan-2020")
)

# Map estimated Artemisinin and Chloroquine phenotypic resistance prevalence for sampleSet "Laos" for both Province and District level for calender years 2018-2020.
mapDrugResistancePrevalence (ctx, sampleSet="Laos", timePeriods=periods,
drugs=c(“Artemisinin”, “Chloroquine”), aggregate=c("Province","District"),
minAggregateCount=10, showNames=TRUE, markerSize=16,
width=15, height=15)

}
\seealso{
Useful links:
\itemize{
\item Details of how drug resistance phenotypes are predicted based on genetic markers {\url{http://ngs.sanger.ac.uk/production/malaria/Resource/29/20200705-GenRe-05-PhenotypeRules-0.39.pdf}}
\item Other useful background information about the SpotMalaria platform  {\url{https://www.malariagen.net/resource/29}}
}
}
