% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapGroupFrequencies}
\alias{mapGroupFrequencies}
\title{Title}
\usage{
mapGroupFrequencies(
  ctx,
  sampleSet,
  type = c("bar", "pie"),
  similarityLevels = 1,
  aggregate = 1,
  minGroupSize = 10,
  showNames = TRUE,
  markerScale = 0.8,
  width = 15,
  height = 15
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
