% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapGroupGraph}
\alias{mapGroupGraph}
\title{Title}
\usage{
mapGroupGraph(
  ctx,
  sampleSet,
  attributes = NULL,
  similarityLevels = 1,
  minGroupSize = 5,
  minConnectionSimilarity = 0.7,
  graphLayout = "fr",
  weighting = "identitySquared",
  width = 15,
  height = 15
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
