% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapGroupPrevalence}
\alias{mapGroupPrevalence}
\title{Title}
\usage{
mapGroupPrevalence(
  ctx,
  sampleSet,
  similarityLevels = 1,
  aggregate = 1,
  minAggregateCount = 5,
  showNames = TRUE,
  minGroupSize = 10,
  markerScale = 0.8,
  width = 15,
  height = 15
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
