% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapGroupPrevalence}
\alias{mapGroupPrevalence}
\title{Map Barcode Group prevalence}
\usage{
mapGroupPrevalence(
  ctx,
  sampleSet,
  aggregate = "Province",
  minAggregateCount = 5,
  showNames = TRUE,
  markerScale = 0.8,
  similarityLevels = 1,
  clusteringMethod = "allNeighbours",
  minGroupSize = 10,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD}

\item{markerScale}{TBD}

\item{similarityLevels}{TBD}

\item{clusteringMethod}{TBD}

\item{minGroupSize}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Map Barcode Group prevalence
}
\examples{
 #TBD
}
