% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapGroupSharing}
\alias{mapGroupSharing}
\title{Map barcode sharing}
\usage{
mapGroupSharing(
  ctx,
  sampleSet,
  type = c("bar", "pie"),
  similarityLevels = 1,
  aggregate = 1,
  minAggregateCount = 5,
  showNames = TRUE,
  markerScale = 0.8,
  minGroupSize = 10,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{type}{TBD}

\item{similarityLevels}{TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD TBD}

\item{markerScale}{TBD}

\item{minGroupSize}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Map barcode sharing
}
\examples{
 #TBD
}
