% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapLocations}
\alias{mapLocations}
\title{Map of Locations}
\usage{
mapLocations(
  ctx,
  sampleSet,
  aggregate = "Province",
  markerSize = c(4, 40),
  showNames = TRUE,
  colourBy = "Province",
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used; must have been previusly created by selectSampleSet().}

\item{aggregate}{The administrative level at which we aggregate (Province/District/Site)}

\item{markerSize}{Allows adjustment of the size of markers on the map. If only one value is passed, 
all markers will be that size; if two values are passed, they will be used as the min 
and max sizeof the marker, whose size will reflect the number of samples.}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District)}

\item{colourBy}{Shows the aggregation level to be used to colour the markers (Country or Province)}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\description{
Creates a map showing markers for all sites/districts/provinces where samples are collected.
For each location, we place a marker on the map, coloured according to the administrative
division it is in
}
\examples{
#TBD
}
