% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapMutationPrevalence}
\alias{mapMutationPrevalence}
\title{Map prevalence of genetic mutations of drug resistance}
\usage{
mapMutationPrevalence(
  ctx,
  sampleSet,
  timePeriods = NULL,
  mutations = "ALL",
  aggregate = "Province",
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used; must have been previously created by selectSampleSet().}

\item{timePeriods}{A separately given list of time period object for partitioning samples into time-interval plots. Each list specifies a time period, with three parameters: name, type and, start. Parameter type can currently only be year. Parameter start must be a date in the format “dd-MMM-yyyy”, and parameter name must be provided. The string passed in the name parameter is added to the end of the file name of the produced files.
Currently, the public GRC data only contains years, so only the year in the provided “start” date is used.}

\item{mutations}{The genetic mutation types for which prevalence will be calculated; "ALL" creates maps for all available genetic mutation types,
which include "kelch13", "arps10", "ferredoxin", "mdr2" (conferring resistance to Artemisinin), "crt" (conferring resistance to Chloroquine),
"dhfr" (conferring resistance to Pyrimethamine), "dhps" (conferring resistance to Sulfadoxine), "plasmepsin 2/3" (conferring resistance to Piperaquine),
"mdr1" (conferring resistance to multiple drugs).
To specify a mutation put the mutation type in between quotation marks e.g. "dhfr", or  c(“arps10”, “kelch13”, “mdr1”) to select several genetic mutation types.}

\item{aggregate}{The administrative level at which we aggregate.}

\item{minAggregateCount}{The minimum count of aggregated samples. To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District)}

\item{markerSize}{Allows adjustment of the size of markers on the map.}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\description{
Creates a map showing the prevalence of genetic mutations known to be associated with antimalarial drug resistance.
If multiple mutation types are specified, then a map will be created for each mutation type.
The prevalence of genetic mutation types are aggregated at the desired administrative level: Province (level 1), or District (level 2). Separate maps are created for each mutation-administrative-combination.
For each aggregation unit, we place a marker on the map, coloured according to the prevalence of the genetic mutation type, with a label indicating the prevalence.
To avoid estimating on very small samples, one can set a minimum count of samples, below which the marker is not shown.
Additionally data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/sampleSetName/map-mutation/.
}
\examples{
#Given lists of time periods of interest, in this case calender years
periods <- list(
list(name="2018", type="year", start="1-Jan-2018"),
list(name="2019", type="year", start="1-Jan-2019"),
list(name="2020", type="year", start="1-Jan-2020")
)

# Produce prevalence maps for all available genetic mutation types, for sampleSet "Laos", for District level, and for calender years 2018-2020.
mapMutationPrevalence (ctx, sampleSet="Laos", timePeriods=periods,
mutations="ALL", aggregate="District",
minAggregateCount=10, showNames=TRUE, markerSize=16,
width=15, height=15)
}
