% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapMutationPrevalence}
\alias{mapMutationPrevalence}
\title{Map the prevalence of genetic mutations.}
\usage{
mapMutationPrevalence(
  ctx,
  sampleSet,
  mutations = "ALL",
  aggregate = 1,
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{mutations}{TBD}

\item{aggregate}{TBD}

\item{minAggregateCount}{TBD}

\item{showNames}{TBD}

\item{markerSize}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Map the prevalence of genetic mutations.
}
\examples{
 #TBD
}
