% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapMutationPrevalence}
\alias{mapMutationPrevalence}
\title{Title}
\usage{
mapMutationPrevalence(
  ctx,
  sampleSet,
  mutation = "ALL",
  aggregate = 1,
  minAggregateCount = 10,
  showNames = TRUE,
  markerSize = 16,
  width = 15,
  height = 15
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
