% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{mapSampleCounts}
\alias{mapSampleCounts}
\title{Map of Sample Counts}
\usage{
mapSampleCounts(
  ctx,
  sampleSet,
  timePeriods = NULL,
  aggregate = "Province",
  minAggregateCount = 1,
  markerSize = c(4, 40),
  markerFontSize = 6,
  colourBy = "Province",
  showNames = TRUE,
  nameFontSize = 5,
  ...
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used, which must have been previously created by selectSampleSet().}

\item{timePeriods}{A separately given list of time period objects for partitioning samples into time-interval plots. Each list specifies a time period, with three parameters: name, type and, start. 
Parameter type can currently only be year. Parameter start must be a date in the format ?dd-MMM-yyyy?, and parameter name must be provided. The string passed in the name parameter is added to the end of the file name of the produced files.
Currently, the public GRC data only contains years, so only the year in the provided ?start? date is used.}

\item{aggregate}{The administrative level at which we aggregate (Province or District). Separate maps are created for each administrative level.}

\item{minAggregateCount}{The minimum count of aggregated samples, below which a marker is not shown.}

\item{markerSize}{Set the size of markers on the map. If only one value is passed, 
all markers will be that size; if two values are passed, they will be used as the min 
and max size of the marker, whose size will reflect the number of samples.}

\item{markerFontSize}{Set the font size of the value labels inside the markers on the map, default=6}

\item{colourBy}{Shows the aggregation level to be used to colour the markers (Country or Province)}

\item{showNames}{If TRUE, labels are shown with the name of the aggregation unit (Province or District)}

\item{nameFontSize}{Set the font size of the geographical name labels outside the markers on the map, if showNames=TRUE, default=5}

\item{...}{any of plot parameters, including: width, height, units, dpi, legendPosition, legendWidth. 
width: the width of the plot (numeric), default=15
height: the height of the plot (numeric), default=15
units: the units in which the width and height are expressed. Supported values: "in" (default), "cm", "mm", "px"
dpi: the resolution of the plot output, expressed as dots per inch, default=300
format: the file format in which the plot will be saved. Supported values: "png" (default), "pdf"
legendPosition: specifies where the legend should be plotted. Supported values: "inset" (default), "separate"
legendWidth: specifies how wide a fixed width legend space should be, default:"NULL"}
}
\description{
Creates maps showing the numbers of samples collected at different aggregation levels.
For each aggregation unit, we place a marker on the map, coloured according to the administrative
division it is in.
Additionally data files are created. Created .tab files can be opened with for example Excel.
Maps and data are located in .../output/out/sampleSetName/map-sampleCount/.
}
\examples{
\dontrun{
# Provide lists of time periods of interest, in this case calendar years
periods <- list(
               list(name="2018", type="year", start="1-Jan-2018"),
               list(name="2019", type="year", start="1-Jan-2019"),
               list(name="2020", type="year", start="1-Jan-2020")
            )

# Create maps showing the numbers of samples collected at Province and District aggregation levels.
mapSampleCounts (ctx, sampleSet="Laos", timePeriods=periods,
                 aggregate=c("Province","District"), minAggregateCount=1,
                 markerSize=c(10,40), colourBy="Province", showNames=TRUE)

# Provide lists of time periods of interest, in this case the Southern Laos malaria season
periods2 <- list(
                list(name="2018", type="year", start="1-Sep-2018"),
                list(name="2019", type="year", start="1-Sep-2019"),
                list(name="2020", type="year", start="1-Sep-2020")
            )

#Maps showing the numbers of samples collected at Province and District aggregation levels.
mapSampleCounts (ctx, sampleSet="Laos", timePeriods=periods2,
                 aggregate=c("Province","District"), minAggregateCount=1,
                 markerSize=c(10,40), colourBy="Province", showNames=TRUE,
                 width=15, height=15, units="in")
}
}
