% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotBarcodeNetwork}
\alias{plotBarcodeNetwork}
\title{Title}
\usage{
plotBarcodeNetwork(
  ctx,
  sampleSet,
  attributes = NULL,
  aggregateCountMin = 10,
  width = 15,
  height = 15
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
