% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotClusterGraph}
\alias{plotClusterGraph}
\title{Plot graphs of bracode identity, using cluster}
\usage{
plotClusterGraph(
  ctx,
  sampleSet,
  clusterSet,
  graphLayout = "fr",
  weightPower = 2,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD}

\item{clusterSet}{TBD}

\item{graphLayout}{TBD}

\item{weightPower}{TBD}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\value{
TBD
}
\description{
Plot graphs of bracode identity, using cluster
}
\examples{
 #TBD
}
