% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotPopulationStructure}
\alias{plotPopulationStructure}
\title{Plot Population structure (PCA and NJT)}
\usage{
plotPopulationStructure(
  ctx,
  sampleSet,
  type = c("njt", "PCoA"),
  attributes = NULL,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD TBD}

\item{type}{TBD}

\item{attributes}{TBD}

\item{width}{TBD}

\item{height}{TBD}
}
\value{
TBD
}
\description{
Plot Population structure (PCA and NJT)
}
\examples{
 #TBD
}
