% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotPrincipalComponents}
\alias{plotPrincipalComponents}
\title{Principal Component Analyses (PCA)}
\usage{
plotPrincipalComponents(
  ctx,
  sampleSet,
  type = "PCoA",
  plots,
  impute = FALSE,
  ...
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext()}

\item{sampleSet}{The name of the sample set being used, which must have been previously created by selectSampleSet()}

\item{type}{The type of method used for the principal component analysis
"PCoA", performs multidimensional scaling on a distance matrix, in which each sample is a variable.
"MCA", performs Multiple Correspondence Analysis on an allele label matrix, in which each sample is a variable.
The default is "PCoA"}

\item{plots}{The list of attributes for which plots will be created.}

\item{impute}{To use imputed or filtered data. The default is TRUE. Ignored for MCA, which always uses imputed data.}

\item{...}{Aesthetics: any of the following plot parameters,
width: the width of the plot (numeric), default=15
height: the height of the plot (numeric), default=15
units: the units in which the width and height are expressed. Supported values: "in" (default), "cm", "mm", "px"
dpi: the resolution of the plot output, expressed as dots per inch, default=300
format: the file format in which the plot will be saved. Supported values: "png" (default), "pdf"
legendPosition: specifies where the legend should be plotted. Supported values: "inset" (default), "separate"
legendWidth: specifies how wide a fixed width legend space should be, default="NULL"
legendDirection: specifies location of legend. Supported values: "vertical" (default), "horizontal"
legendFontSize: specifies font size of the legend (numeric), default=4
axisTitleSize: specifies axis label font size (numeric), default=1}
}
\value{
Principal component plots and .tab data files.
}
\description{
Principal Component Analysis (PCA), Principal Coordinate Analysis (PCoA) and Multiple Correspondence Analysis (MCA) are methods of reducing complexity of the data
(in this case, a matrix of genetic barcodes genotyped across a number of variants) down to fewer dimensions, while preserving as much information contained 
in the original data as possible. This is done by deriving mathematically a set of principal components that explain most of the observed variance in the dataset.
grcMalaria does not implement a "true PCA" methods, since this would require assigning numeric scalar values to alleles, which is problematic for variant that 
have more than two alleles (for example, GenRe-Mekong microhaplotype barcode data). Instead, we implement the following:
- PCoA (default method), which analyzes a pairwise distance matrix between all samples, computed from the barcode allele data;
- MCA which analyzes the barcode allele data, treating them as categorical labels; in this case, the data must contain exactly one allele per sample at 
  each variant (hence, MCA uses the imputed dataset).
One plot is created for each specified graphical attribute in the parameter 'plots', in which the samples are coloured according to user input.
Graphical display of attributes needs to be specified before using this function (see function loadGraphicAttributes).
The PCs (Principal Components) are arranged in descending order of explained variance, with PC1 axis showing the largest variation and PC2 axis showing the second 
largest variation and so on.
}
\examples{
\dontrun{
# Perform a principal component analysis for province, kelch13, plasmepsin2/3 amplification
# status, and kelch13 in combination with plasmepsin2/3 amplification status,
# using the "PCoA" method on sampleSet "Laos"
plotPrincipalComponents (ctx, sampleSet="Laos", type="PCoA",
    plots=list(
              list(name="ByProvince",       attributes="province"),
              list(name="ByKelch13",        attributes="k13"),
              list(name="ByPm23",           attributes="pm23"),
              list(name="ByKelch13AndPm23", attributes=c("k13","pm23-noColour"))
           ),
    width=15, height=10)
}
}
