% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotPrincipalComponents}
\alias{plotPrincipalComponents}
\title{Plot population structure (PCA)}
\usage{
plotPrincipalComponents(
  ctx,
  sampleSet,
  type = "PCoA",
  plots,
  width = 15,
  height = 15
)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used; must have been previously created by selectSampleSet().}

\item{type}{The type of method used for the principal component analysis.
"PCoA", performs multidimensional scaling on a distance matrix, in which each sample is a variable.
"nipals", "bpca" are methods applied to the set of genetic barcode genotypes (each barcode position is a variable).
The default is "PCoA".}

\item{plots}{The list of attributes for which plots will be created. See example below.}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\value{
Principal component plots and .tab data files.
}
\description{
This function produces principal component analysis (PCA) plots, in which the population (i.e. samples in the sampleSet),
is structured in principal components based on their 'genetic barcode'.
The 'genetic barcode' consists of 101 SNPs determined for each sample. These SNPs are distributed across all nuclear chromosomes.
SNP positions were chosen based on their ability to differentiate populations and their power to recapitulate genetic distance.
One plot is created for each specified attribute in the parameter 'plots', in which the samples are coloured according to user input.
Graphical display of attributes needs to be specified before using this function.
Give command ?loadGraphicAttributes for more information.
Additionally, data files are created. Created .tab files can be opened with for example Excel.
Plots and data are located in .../output/out/(sampleSetName)/(PCA-method, see parameter 'type')/.
}
\examples{
# Perform a principal component analysis for province, kelch13, plasmepsin 2,3, and kelch13 in combination with plasmepsin 2,3,
# using the "PCoA" method on sampleSet "Laos"
plotPrincipalComponents (ctx, sampleSet="Laos", type="PCoA",
plots=list(
 list(name="ByProvince",       attributes="province"),
 list(name="ByKelch13",        attributes="k13"),
 list(name="ByPm23",           attributes="pm23"),
 list(name="ByKelch13AndPm23", attributes=c("k13","pm23-noColour"))
), width=15, height=10)

}
