% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotTree}
\alias{plotTree}
\title{Plot Trees}
\usage{
plotTree(ctx, sampleSet, type = "njt", plots, impute = FALSE, ...)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSet}{The name of the sample set being used, which must have been previously created by selectSampleSet().}

\item{type}{The method used to create the tree.Currently the only option "njt" which creates a neighbour-joining tree.}

\item{plots}{The list of attributes for which plots will be created. See example below.}

\item{impute}{To use imputed or filtered data. The default is TRUE.}

\item{...}{Aesthetics: any of the following plot parameters,
width: the width of the plot (numeric), default=15
height: the height of the plot (numeric), default=15
units: the units in which the width and height are expressed. Supported values: "in" (default), "cm", "mm", "px"
dpi: the resolution of the plot output, expressed as dots per inch, default=300
format: the file format in which the plot will be saved. Supported values: "png" (default), "pdf"
legendPosition: specifies where the legend should be plotted. Supported values: "inset" (default), "separate"
legendWidth: specifies how wide a fixed width legend space should be, default="NULL"
legendDirection: specifies location of legend. Supported values: "vertical" (default), "horizontal"
legendFontSize: specifies font size of the legend (numeric), default=4
axisTitleSize: specifies axis label font size (numeric), default=1}
}
\value{
Phylogenetic tree plots, .tab data files, and a .newick tree file.
}
\description{
Neighbor-joining (NJ) tree is a another method to show how samples are related based on the distances between barcodes.
It is not a phylogenetic tree. The NJ tree is unrooted, which means: 1) It clusters together related sequences using the distance matrix,
the tree has no orientation, there is no assumption about ancestry.
For each specified attribute, a plot is created for the phylogenetic tree, in which the nodes (i.e. samples) are coloured according to that attribute.
Graphical display of attributes needs to be specified before using this function.
In addition to the plot, a newick file is generated, which can be used to upload in tree visualisation tools (for example https://microreact.org/)
}
\examples{
\dontrun{
# Produce a neighbour-joining tree for sampleSet "Laos"
# with plots in which the samples (i.e. nodes) are coloured according to i) province, ii) kelch13,
# iii) plasmepsin2/3 amplification status, iv) kelch13 and plasmepsin2/3 amplification status
plotTree (ctx, sampleSet="Laos", type="njt",
          plots=list(
                    list(name="ByProvince",       attributes="province"),
                    list(name="ByKelch13",        attributes="k13"),
                    list(name="ByPm23",           attributes="pm23"),
                    list(name="ByKelch13AndPm23", attributes=c("k13","pm23-noColour"))
          ),
          width=15, height=10)
}
}
