% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotTree}
\alias{plotTree}
\title{Plot a tree (e.g. neighbour joining tree, NJT)}
\usage{
plotTree(
  ctx,
  sampleSet,
  method = "njt",
  attributes = NULL,
  width = 15,
  height = 15,
  colouredBranches = TRUE,
  showLeaves = TRUE,
  nonLeafColour = "darkgrey",
  branchThickness = 3,
  popLabels = FALSE,
  sampleLabels = FALSE
)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD TBD}

\item{attributes}{TBD}

\item{width}{TBD}

\item{height}{TBD}

\item{colouredBranches}{TBD}

\item{showLeaves}{TBD}

\item{nonLeafColour}{TBD}

\item{branchThickness}{TBD}

\item{popLabels}{TBD}

\item{sampleLabels}{TBD}

\item{type}{TBD}
}
\value{
TBD
}
\description{
Plot a tree (e.g. neighbour joining tree, NJT)
}
\examples{
 #TBD
}
