% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{plotTree}
\alias{plotTree}
\title{Plot Trees}
\usage{
plotTree(ctx, sampleSet, type = "njt", plots, width = 15, height = 15)
}
\arguments{
\item{ctx}{TBD}

\item{sampleSet}{TBD TBD}

\item{type}{TBD}

\item{plots}{TBD}

\item{width}{The width (in inches) of the map image.}

\item{height}{The height (in inches) of the map image.}
}
\value{
TBD
}
\description{
Plot Trees
}
\examples{
 #TBD
}
