% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{resetColourPalette}
\alias{resetColourPalette}
\title{Reset to the default colour palette, removing user-defined palettes}
\usage{
resetColourPalette(ctx)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}
}
\value{
Return to the default colour palette of grcMalaria.
}
\description{
Return to the default 25-colour palette of grcMalaria.
}
\examples{
\dontrun{
# return to the default 25-colour palette:
resetColourPalette(ctx)

# To view default 25-colour palette
ctx$config$defaultPalette
}
}
