% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{selectSampleSet}
\alias{selectSampleSet}
\title{Select a set of samples using metadata}
\usage{
selectSampleSet(ctx, sampleSetName, select)
}
\arguments{
\item{ctx}{The analysis context, created by intializeContext().}

\item{sampleSetName}{Select a name for the sample set. You will use this for subsequent analysis tasks. This will create a folder with the same name e.g. “Laos”.}

\item{select}{The criteria for sample selection. Selected samples will match all specified criteria. The values must be a list. "field" corresponds to a column name in the datafile. "values" correspond to an array of possible values that can be matched for a sample to be selected. Use "," to add additional parameters inside a quotation mark " ".}
}
\value{
The analysis context, augmented with the new sample set
}
\description{
Selects a set of sample for analysis, based on their metadata values.
A given analysis context can contain multiple sample sets. These can be labelled with different names.
The selection criteria are specified as a list containing a sequence of lists with two elements each:
"field" which is the column name to be checked, and "values" which is an array of possible values that can be matched for a sample to be selected.
A selected sample must match all the criteria.
Repeat this step for the different sample sets you would like to analyze
}
\examples{
\dontrun{
## Select a sample set to work on ##

# To select samples from 1 field (1 column in the GRC)
selectSampleSet(ctx, sampleSetName="EBKK", select=list(
                                list(field="Country", values=c("VN", "KH", "LA")) ))

# To select samples from 2 fields
selectSampleSet(ctx, sampleSetName="Laos", select=list(
                                list(field="TimePoint", values=c("D00H00","-")),
                                list(field="Study", values=c("1208-PF-LA-CMPE-GENRE")) ))

# To select samples from 3 fields
selectSampleSet(ctx, sampleSetName="SouthLA_2017", select=list(
                                list(field="Country", values="LA"),
                                list(field="AdmDiv1", values=c("Attapeu", "Champasak")),
                                list(field="Year", values=c("2017", "2018")) ))

}
}
